+++ asistente_ladm_col/asistente_ladm_col_plugin.py
@@ -64,7 +64,7 @@ from .gui.create_boundaries_cadastre_wizard import CreateBoundariesCadastreWizar
 from .gui.create_building_cadastre_wizard import CreateBuildingCadastreWizard
 from .gui.create_building_unit_cadastre_wizard import CreateBuildingUnitCadastreWizard
 from .gui.create_right_of_way_cadastre_wizard import CreateRightOfWayCadastreWizard
-from .gui.associate_extaddress_cadastre_wizard import AssociateExtAddressWizard
+from .gui.wizards.cadastre.associate_extaddress_cadastre_wizard import AssociateExtAddressWizard
 from .gui.create_col_party_cadastre_wizard import CreateColPartyCadastreWizard
 from .gui.create_group_party_cadastre import CreateGroupPartyCadastre
 from .gui.create_legal_party_prc import CreateLegalPartyPRCWizard
